move test code to unit test
